更新页面图片数据
